cleaned up db store and get tests, changed uint to int to make it run

+++ src/Model/DatabaseModule/Database.cs
@@ -323,9 +323,9 @@ internal class Database : IDatabase {
         return result;
     }
 
-    public uint GetNextUserId() {
+    public int GetNextUserId() {
         Guid guid = Guid.NewGuid();
-        uint result = (uint) guid.GetHashCode();
+        int result = guid.GetHashCode();
         return result;
     }
 }
